requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.2
openai==1.82.1
python-dotenv==1.0.1
jinja2==3.1.3
weasyprint==61.2
pdfkit==1.0.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
requests==2.31.0
pydantic==2.7.1